hw13_2/pyproject.toml
[tool.poetry]
name = "hw10"
version = "0.1.0"
description = ""
authors = ["PollyK21 <pol.kulish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
pymongo = "^4.6.3"
mongoengine = "^0.28.2"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
